Access Wifi Keystore HAL via binder
The test originally linked against the HAL directly, as well as the
generated stub to access the HAL.. This "worked", however if both the
32-bit and 64-bit versions of the generated stub library were
not present the test would fail. It turned out that on AOSP builds
this was the case, as a 32-bit version of android.system.wifi.keystore
was not present as it was never used at runtime.

To avoid this issue, the stub is now linked statically. Further, the
HAL under test is accessed over binder.

Bug: 130207625
Test: On an AOSP build:
      make vts;
      vts-tradefed run commandAndExit vts -m VtsHalWifiKeystoreV1_0TargetTest

Change-Id: I02cac1e89c8bed45a4b9b98820c86500327891b6
Merged-In: I138f3e7809a337cec81dccbc52c7a9170acd1d3d