Split Wifi Keystore HAL tests into smaller pieces
There was originally one test per function, which attempted many
possible error conditions and one success condition. It would be
easier to reason through the tests and maintain them if they
were smaller and each test attempted on thing.

Note that the logic on the "wrong user" tests was reversed; the
HAL expects the user to be the wifi user. The reverse may have
originally worked because the HAL library was run within the test
process.

Bug: 130207625
Test: On an AOSP build:
      make VtsHalWifiKeystoreV1_0TargetTest;
      vts-tradefed run commandAndExit vts -m VtsHalWifiKeystoreV1_0TargetTest

Change-Id: I534b7f66cdb1047a972a4dab708609b15cf287df
Merged-In: I282e151d0b38903578bd9077814c810e44b3b93d