Relax testing under GSI
VTS testing on GSI images will no longer occur on userdebug
builds. Instead, a user build is used with a boot image containing
a rootable adb. There is a change in Keystore that allows root to
create test keys on debuggable builds specifically for this VTS test.
Without that change the positive test cases will fail, as the test
keys cannot be generated.

To allow this test to continue working when run against GSI builds,
the parts of the test which require debuggable builds are skipped.

Test: build/test VtsHalWifiKeystoreV1_0TargetTest
      on a GSI user build with rootable adb
Bug: 135421408

Change-Id: Ie52e916a658b09eea66ffd9c5f2ac704656f4f3f