Keystore 2.0: Revise AIDL interface.
This patch removes all output parameters by moving results into
parcelables designated to be return values.

We also introduced KeyFlags. Specifically, we introduced
AUTH_BOUND_WITHOUT_CRYPTOGRAPHIC_LSKF_BINDING. This flag allows the
creation of keys that are logically authentication bound but are not
encrypted with the master key. Keys like this are used to create a
binding between the synthetic password derivation and a hardware bound
key and is used when weaver is not available.

Test: VTS test in development
Change-Id: Iad93092eaeedba41eaabde58c6dabda7f181f7da